refactor(release): split candidate orchestration boundaries
Keep Runner as the stable facade while separating repository, cache, gate, local-attempt, and remote-run ownership. Capture hosted dogfood evidence and repair GitHub identical-compare and targeted-retry identity handling.
