+++ test/unit/release_contract_test.rb
@@ -363,6 +363,9 @@ class ReleaseContractTest < Minitest::Test
     body = File.read(CANDIDATE_WORKFLOW)
     workflow = YAML.safe_load_file(CANDIDATE_WORKFLOW, aliases: true)
     jobs = workflow.fetch("jobs")
+    validate_body = jobs.fetch("validate").fetch("steps").find do |step|
+      step["id"] == "identity"
+    end.fetch("run")
 
     assert_includes body, "selected_gates"
     assert_includes body, "RetrySelection"
@@ -373,6 +376,10 @@ class ReleaseContractTest < Minitest::Test
     assert_includes body, '.external_id == ("hive-release-candidate:v1:"'
     assert_includes body, "display_title"
     refute_includes body, "|| true"
+    assert_includes validate_body,
+                    "check-runs?check_name=hive-release-candidate&filter=all&per_page=100"
+    assert_includes validate_body, '.head_sha == $candidate'
+    refute_includes validate_body, '.details_url'
 
     %w[catalog release-e2e package managed-web freshness candidate-version].each do |job_name|
       job = jobs.fetch(job_name)
